services/gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
pydantic-settings==2.1.0
starlette==0.27.0